api_project_2022/requirements.txt
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
idna==2.10
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
PyYAML==5.4.1
requests==2.25.1
toml==0.10.2
urllib3==1.26.6
